+++ src/main/java/com/sparta/parknav/_global/data/KakaoMapService.java
@@ -1,4 +1,4 @@
-package com.sparta.parknav.global.data;
+package com.sparta.parknav._global.data;
 
 import com.sparta.parknav.parking.dto.KakaoSearchDto;
 import org.json.JSONArray;
