+++ src/main/java/com/sparta/parknav/_global/response/MsgType.java
@@ -1,4 +1,4 @@
-package com.sparta.parknav.global.response;
+package com.sparta.parknav._global.response;
 
 import lombok.Getter;
