+++ src/main/java/com/sparta/parknav/_global/response/ResponseUtils.java
@@ -1,6 +1,6 @@
-package com.sparta.parknav.global.response;
+package com.sparta.parknav._global.response;
 
-import com.sparta.parknav.global.exception.ErrorResponse;
+import com.sparta.parknav._global.exception.ErrorResponse;
 
 public class ResponseUtils {
