+++ src/main/java/com/sparta/parknav/booking/controller/BookingController.java
@@ -4,8 +4,8 @@ import com.sparta.parknav.booking.service.BookingService;
 import com.sparta.parknav.booking.dto.BookingInfoRequestDto;
 import com.sparta.parknav.booking.dto.BookingInfoResponseDto;
 import com.sparta.parknav.booking.dto.BookingResponseDto;
-import com.sparta.parknav.global.response.ApiResponseDto;
-import com.sparta.parknav.global.security.UserDetailsImpl;
+import com.sparta.parknav._global.response.ApiResponseDto;
+import com.sparta.parknav._global.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
