+++ src/main/java/com/sparta/parknav/booking/service/BookingService.java
@@ -7,11 +7,11 @@ import com.sparta.parknav.booking.entity.Car;
 import com.sparta.parknav.booking.entity.ParkBookingInfo;
 import com.sparta.parknav.booking.repository.CarRepository;
 import com.sparta.parknav.booking.repository.ParkBookingInfoRepository;
-import com.sparta.parknav.global.exception.CustomException;
-import com.sparta.parknav.global.exception.ErrorType;
-import com.sparta.parknav.global.response.ApiResponseDto;
-import com.sparta.parknav.global.response.MsgType;
-import com.sparta.parknav.global.response.ResponseUtils;
+import com.sparta.parknav._global.exception.CustomException;
+import com.sparta.parknav._global.exception.ErrorType;
+import com.sparta.parknav._global.response.ApiResponseDto;
+import com.sparta.parknav._global.response.MsgType;
+import com.sparta.parknav._global.response.ResponseUtils;
 import com.sparta.parknav.management.service.ParkingFeeCalculator;
 import com.sparta.parknav.parking.entity.ParkInfo;
 import com.sparta.parknav.parking.entity.ParkOperInfo;
