+++ src/main/java/com/sparta/parknav/parking/controller/ParkSearchController.java
@@ -1,6 +1,6 @@
 package com.sparta.parknav.parking.controller;
 
-import com.sparta.parknav.global.response.ApiResponseDto;
+import com.sparta.parknav._global.response.ApiResponseDto;
 import com.sparta.parknav.parking.dto.ParkSearchResponseDto;
 import com.sparta.parknav.parking.dto.ParkSearchRequestDto;
 import com.sparta.parknav.parking.service.ParkSearchService;
