+++ src/main/java/com/sparta/parknav/user/service/UserService.java
@@ -1,11 +1,11 @@
 package com.sparta.parknav.user.service;
 
-import com.sparta.parknav.global.exception.CustomException;
-import com.sparta.parknav.global.exception.ErrorType;
-import com.sparta.parknav.global.jwt.JwtUtil;
-import com.sparta.parknav.global.response.ApiResponseDto;
-import com.sparta.parknav.global.response.MsgType;
-import com.sparta.parknav.global.response.ResponseUtils;
+import com.sparta.parknav._global.exception.CustomException;
+import com.sparta.parknav._global.exception.ErrorType;
+import com.sparta.parknav._global.jwt.JwtUtil;
+import com.sparta.parknav._global.response.ApiResponseDto;
+import com.sparta.parknav._global.response.MsgType;
+import com.sparta.parknav._global.response.ResponseUtils;
 import com.sparta.parknav.user.dto.LoginRequestDto;
 import com.sparta.parknav.user.dto.SignupRequestDto;
 import com.sparta.parknav.user.entity.Admin;
